azure_functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
gremlinpython
apify-client
python-dotenv
pydantic
requests
openai[datalib]
pytest